+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Avoid print server config users in console help
+
 ### 1.0.0 (2014-08-25)
 
 * Initial release